tela de inserção de dados

+++ MauiAppMInhasCompras/Models/Produto.cs
@@ -7,7 +7,7 @@ namespace MauiAppMInhasCompras.Models
         [PrimaryKey, AutoIncrement]
         public int Id { get; set; }
         public string Descricao { get; set; }
-        public decimal Preco { get; set; }
-        public int Quantidade { get; set; }
+        public double Preco { get; set; }
+        public double Quantidade { get; set; }
     }
 }
